+++ packages/canonical-bridge-widget/src/modules/aggregator/adapters/cBridge/components/CBridgeOption.tsx
@@ -81,7 +81,7 @@ export const CBridgeOption = () => {
       isError={isError}
       onSelectBridge={onSelectBridge}
     >
-      <RouteName isError={isError} bridgeType="cBridge" />
+      <RouteName isError={isError} bridgeType="cBridge" receiveAmt={receiveAmt} />
       <RouteTitle
         isError={isError}
         receiveAmt={receiveAmt}