+++ packages/canonical-bridge-widget/src/modules/transfer/components/TransferOverview/RouteInfo/RouteName/index.tsx
@@ -9,15 +9,17 @@ import { useGetBestTime } from '@/modules/transfer/hooks/useGetBestTime';
 import { getMaxValueKey } from '@/core/utils/number';
 import { BestRouteTag } from '@/modules/transfer/components/TransferOverview/RouteInfo/RouteName/BestRouteTag';
 import { useBridgeConfig } from '@/CanonicalBridgeProvider';
+import { RedeemRatio } from '@/modules/transfer/components/TransferOverview/RouteInfo/RouteName/RedeemRatio.tsx';
 
 interface RouteNameProps {
   bridgeType?: BridgeType;
   isReceiveSection?: boolean;
   isError?: boolean;
+  receiveAmt?: string;
 }
 
 export const RouteName = React.memo(
-  ({ bridgeType, isReceiveSection = false, isError }: RouteNameProps) => {
+  ({ bridgeType, isReceiveSection = false, isError, receiveAmt }: RouteNameProps) => {
     const { formatMessage } = useIntl();
     const { getSortedReceiveAmount } = useGetReceiveAmount();
     const bridgeConfig = useBridgeConfig();
@@ -31,9 +33,12 @@ export const RouteName = React.memo(
         className="bccb-widget-route-name"
         minH={'24px'}
         flexDir={'row'}
-        gap={'8px'}
+        gap={'4px 8px'}
         alignItems={'center'}
         opacity={isError ? 0.5 : 1}
+        flexWrap={'wrap'}
+        whiteSpace={'nowrap'}
+        mb={isReceiveSection ? '4px' : 0}
       >
         {bridgeType === 'cBridge' ? (
           <CBridgeIcon w={['16px', '16px', '16px', '20px']} h={['16px', '16px', '16px', '20px']} />
@@ -74,7 +79,7 @@ export const RouteName = React.memo(
           ]}
           gap={'8px'}
           alignItems={'center'}
-          flex={1}
+          flex={[0, 0, 0, 1]}
         >
           <Box
             className="bccb-widget-route-name-text"
@@ -94,10 +99,11 @@ export const RouteName = React.memo(
               ? formatMessage({ id: 'route.option.meson.title' })
               : null}
           </Box>
-          <Flex gap={'4px'}>
-            {bridgeType === bestTimeRoute && <BestRouteTag bestMode={'bestTime'} />}
-            {bridgeType === bestReturnRoute && <BestRouteTag bestMode={'bestReturn'} />}
-          </Flex>
+        </Flex>
+        <Flex gap={'4px'} pr={['48px', '48px', 0]}>
+          {bridgeType === bestTimeRoute && <BestRouteTag bestMode={'bestTime'} />}
+          {bridgeType === bestReturnRoute && <BestRouteTag bestMode={'bestReturn'} />}
+          {!(isError && receiveAmt === '--') && <RedeemRatio bridgeType={bridgeType} />}
         </Flex>
       </Flex>
     );